+++ src/App.js
@@ -1,34 +1,15 @@
 import React from 'react';
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
-const todoList = [
-  { id: 1, 
-    title: "Add To List"
-  },
-  { id: 2, 
-    title: "remove From List"
-  },
-  { id: 3, 
-    title: "completed Assignment"
-  },
-];
 
 
 function App() {
   return (
     <div className="App">
-      <header>
         <h1>Todo List</h1>
-        <ul>
-            {todoList.map((todoList)=> {
-              return (
-                <li> 
-                  {todoList.key}
-                  {todoList.title}
-              </li>
-              )
-            })}
-        </ul>
-        </header>
+        <AddTodoForm/>
+        <TodoList/>
     </div>
   );
 }
